reorganized the test/examples directory structure
