+++ src/test/java/com/attensa/rubberband/Cat.java
@@ -1,4 +1,4 @@
-package com.attensa.rubberband.examples;
+package com.attensa.rubberband;
 
 import lombok.Value;
 import lombok.experimental.Wither;
